use double for the simulation clock

+++ src/luposdate3000_simulator_core/src/commonMain/kotlin/lupos/simulator_core/Entity.kt
@@ -20,13 +20,13 @@ public abstract class Entity : ISimulationLifeCycle {
         }
     }
 
-    protected fun scheduleEvent(destination: Entity, data: Any, delay: Long) {
+    protected fun scheduleEvent(destination: Entity, data: Any, delay: Double) {
         require(!isTerminated)
         val sim = simulation
         sim.addEvent(delay, this, destination, data)
     }
 
-    public fun setTimer(time: Long, callback: ITimer) {
+    public fun setTimer(time: Double, callback: ITimer) {
         scheduleEvent(this, callback, time)
     }
